redesign build process.
We abandon the usage of wasm-builder entirely. Actors are now
built as ordinary Rust libraries when cargo-built individually.

The Wasm bytecode generation is now performed by the bundle module.
This module uses a brand new bundler utility to generate a CARv1 file
containing the Wasm bytecode of all actors, as well as a top-level
index structure associating actor names with their respective CodeCIDs.