fix: build wasm modules in a way that works with crates.io
Previously, we'd build all actors from the "bundler". Unfortunately,
this doesn't work when uploaded to crates.io because the bundler doesn't
have access to the dependencies. In theory, we probably could have made
this work, but reverting to the "substrate-wasm-builder" build style was
simpler.

We chose _not_ to use the substrate wasm builder because:

1. This gives us a bit more control.
2. It works on stable rust.
3. It removes a bunch of dependencies.