build: re-run bundle build if workspace actors have changed (#166)
Rerun if the source, dependencies, build options, build script _or_
actors have changed. We need to check if the actors have changed because
otherwise, when building in a workspace, we won't re-run the build
script and therefore won't re-compile them.

This _isn't_ an issue when building as a dependency fetched from
crates.io (because the crate is immutable).