test runtime: use Rc<MemoryBlockstore>.
With the real FvmRuntime, we expect that cloning the &Blockstore
returned by Runtime#store() would simply clone the reference.

This is because inside the FVM, blockstore calls proxy out to
FVM syscalls; the proxy itself holds no state, and cloning it is
unimportant.

However, in the MockRuntime, we were using a physical MemoryBlockstore.
Cloning it would return a new physical copy, thus breaking the expectation.

This commit makes MockRuntime use an Rc<MemoryBlockstore>, so all clones
will point to the same MemoryBlockstore, thus preserving the expectation.