re-implement SELFDESTRUCT as an execution halt + deletion outside transaction.
We are executing the contract logic in one long-running Runtime transaction.

Thus, SELFDESTRUCT cannot call rt.delete_actor(), as this needs to happen outside
of a transaction for safety reasons.

Furthermore, we were not halting EVM execution on SELFDESTRUCT.

This commit reimplements the way that SELFDESTRUCT is handled by flagging
the deletion in the ExecutionState, and halting execution.

The FVM deletion now happens in a deferred fashion outside the managed
transaction.