EVM runtime: implement GetBytecode and GetStorageAt getters.
GetBytecode returns the CID of the bytecode block. In the future,
the FVM's reachability analysis should add this CID to the reachable set
of the caller. GetBytecode is publicly callable.

GetStorageAt returns the value stored in the supplied EVM storage key.
This method is private and only intended for calls from instrumentation and
client functionality, such as the eth_getStorageAt JSON-RPC API operation.
In order to make it private, it validates that the caller is the zero address.
It needs to be private to preserve EVM guarantees: no contract is able to
arbitrarily access storage keys from another contract. The client must form
a message from the 0x0 sender address for this method to work.