fix: encode/decode EVM parameters as CBOR (#759)
* fix: encode/decode EVM parameters as CBOR

The previous code was passing parameters as raw bytes, but telling the
system that these bytes were valid CBOR. Unfortunately:

- This is incorrect and will break once we actually start enforcing this.
- This makes it _very_ difficult to test the EVM because we have quite a
  bit of testing code that assumes CBOR.

The long-term solution is to correctly handle IPLD (see
https://github.com/filecoin-project/builtin-actors/issues/758). But the
short-term solution is to just encode/decode parameters as CBOR.

* fix: only use RawBytes for CBOR

"RawBytes" is for storing serialized CBOR (per the docs). Any other uses
are bugs.