fix: correctly handle out-of-bounds cases when copying memory
- CALL: truncates the return value, but doesn't zero fill.
- CALLDATA, EXTCODECOPY, CODECOPY: treat input as if it were followed by
  infinite zeros.
- RETURNDATACOPY: explicitly forbids out-of-bounds reads.

We might be able to change the behavior of EXTCODECOPY and CODECOPY to
match RETURNDATACOPY, but we _can't_ change CALLDATA as solidity abuses
this feature for zeroing memory. So we're just going to match what the
EVM does because it's safer (and because we need to implement that
behavior _anyways_ for CALLDATA).

fixes https://github.com/filecoin-project/ref-fvm/issues/1021
fixes https://github.com/filecoin-project/ref-fvm/issues/1024