fix: reject (f4) actor redeploy after selfdestruct
We could also reject this in the FVM, but then we'd have to tell the FVM
that we're re-assigning an existing ID address.

The downside of this method is that it's _slightly_ more expensive (we
have to make multiple syscalls instead of one).

The benefit is that it makes auditing and testing this code _much_
simpler. Before, we were enforcing half the constraints in the FVM, and
the other half in the init actor, which is why we had into this bug in
the first place.

fixes https://github.com/filecoin-project/ref-fvm/issues/1175