EVM: abort if we fail to transfer to the beneficiary on selfdestruct
This is somewhat of a damned if we do, damned if we don't situation.
This operation _can't_ fail in the EVM, but it _can_ fail in the FVM
because we implement this with an actual send.

However, this is far from the only instruction that can fail
unexpectedly (e.g., EXTCODEHASH).

🤷‍♂️

fixes #1558