Simplify pledge update on sector upgrade
This is a small pre-factor for FIP0056 because I'm handling the pledge
the same way there.

Basically, instead of checking if we have enough funds to cover the
pledge each time we update a sector, do so once at the end (and check
the balance invariants). This will save a _bit_ of gas, but, more
importantly, it avoids using two different "pledge deltas".