fix: update (and use) the power-base epoch for non-simple qap
Given:

    | A | B | C |
    |   |   |   + new expiration
    |   |   + old expiration
    |   + new power base (current epoch)
    +- old power base (usually the activation epoch)

Previously, we'd take the remaining weight from B and then normalize it
over A + B + C (new_expiration - activation).

This isn't quite correct, but it was as correct as we could get
without recording the power base epoch. However, now that we have the
power base epoch, we should normalize it (spread it out over) B + C, not
A.

In practice, this just means:

1. Updating the power base when extending a non-simple QaP sector.
2. Adjusting the tests to account for these changes.