Implement SDM
Implement SDM ([FIP0057][]). This builds on #1229 and uses the new
"power base epoch" as the base epoch for computing the sector duration
multiplier.

There are three main components of this PR:

1. Apply the SDM.
2. Update the initial pledge when we extend a sector. I've combined both
extension functions (simple and non-simple) as they now share quite a
bit of logic.
3. Record the extension-based pledge updates in the state.

One tricky part to note is that the power can now change multiple times
over the lifetime of a sector, instead of just once. This makes
termination penalties tricky as there should, in theory, be _multiple_
"replaced day rewards".

To simplify this situation, we:

1. Always update the "expected day reward" based on the latest power.
2. Always update the "replaced day rewards" to the maximum of the old
power and the old replaced day rewards.

We do this for non-simple qap in addition to simple qap.

Finally, unlike the FIP (to be updated), this PR starts applying SDM
from `540 * EPOCHS_IN_DAY` instead of `3/2 * EPOCHS_IN_YEAR` to match
the current maximum sector duration extension. However, it keeps the
same target 2x maximum multiplier.

[FIP0057]: https://github.com/filecoin-project/FIPs/blob/master/FIPS/fip-0057.md