fix: sector activation epoch (#1229)
* Miner actor: fix sector activation epoch

* update expected and replaced day rewards for extensions

* fix: set power_base_epoch in some tests

* nit: max then clone

* test: add power base predicates

* test: fix expected invocations

* fix: update non-verified deal weights when changing the power base

This shouldn't actually affect power calculations (non-verified deals
have the same weight as CC), but it wasn't quite correct either.

* fix: update (and use) the power-base epoch for non-simple qap

Given:

    | A | B | C |
    |   |   |   + new expiration
    |   |   + old expiration
    |   + new power base (current epoch)
    +- old power base (usually the activation epoch)

Previously, we'd take the remaining weight from B and then normalize it
over A + B + C (new_expiration - activation).

This isn't quite correct, but it was as correct as we could get
without recording the power base epoch. However, now that we have the
power base epoch, we should normalize it (spread it out over) B + C, not
A.

In practice, this just means:

1. Updating the power base when extending a non-simple QaP sector.
2. Adjusting the tests to account for these changes.

* fix: calculate the correct unverified deal weight

I flubbed this last time. It doesn't actually matter (not sure why we
still track this) but we might as well get it right.

* fix: take the max of the expected day reward

* Address rebase conflicts

* Address review

* fix: take the max of expected pledges on snap

This should never go down, but we should check anyways.

---------

Co-authored-by: Steven Allen <steven@stebalien.com>