chore: update dependencies (#1599)
* chore: update bundler

This updates the actor bundler to remove the old IPLD libraries (cid
v0.10, etc.) from the dependency tree.

* chore: update pretty_env_logger

Removes dependency on outdated isatty library.

* chore: update test-case dep

This removes our dependency on proc-macro-error

* chore: update url & hashbrown

This removes two more audit warnings.

* chore: update everything else

Why not. This also gets rid of a few dependencies from our tree.

* chore: downgrade cargo-platform to avoid updating the rust compiler

* chore: update deny.toml to account for the dependency updates

* chore: incompatible upgrades

Runtime deps:

- unsigned-varint
- integer-encoding
- num-derive
- itertools
- indexmap
- derive_builder

Dev deps:

- hex_literal
- rlp
- base64

* chore: downgrade rlp

We only depend on this in our tests and we might as well stick to one version.