Use docker environment to build the actors reproducibly (#1606)
* Basic docker build environment.

* Dockerfile: Run scripts by default.

* Add makefile rule for dockerized build.

* Tweak Dockerfile

* Add make *-repro commands for all networks

* Switch to reproducible build in github actions

* Remove -it from reproducible docker run commands

* git ignore everything in output

* Update rust image and specify a index digest.

* remove the deps-build target

* don't try to install anything else

We should try to use the exact image from dockerhub. This means updating
rust will require two steps, but that's not a huge deal IMO.

* do install the rust toolchain before copying

This lets us take advantage of docker layers.

* Make a clean checkout of the repo

This means it isn't possible to reproducibly build a dirty repo but...
nobody wants to do that anyways. It does mean that the reproducible
build won't be affected by other files in the tree.

* Minor improvements to repro builds with Docker (#1607)

* exclude `target` (mine was >40G when I first ran this, not needed for a container build)
* extract common variables in Makefile
* restore missing `all-bundles` target
* make `.PHONY` readable

* Chown after the fact instead of adding a user to the container

This works with both rootless and root docker/podman modes. Otherwise,
when running in rootless mode, the outer user's UID gets mapped to a
temporary UID on the outside, leading to permissions issues when writing
to the output directory.

---------

Co-authored-by: David Himmelstrup <lemmih@gmail.com>
Co-authored-by: Ian Davis <jungziege@gmail.com>
Co-authored-by: Rod Vagg <rod@vagg.org>