build: don't build the bundle from the build script
With this change, we no longer build the bundle with the build script,
exporting it as a global constant. Instead, we require users to either
download the bundle from GitHub or build by executing the root
package.

This should reduce the slowness usually involved in running `cargo
check --all` on this repo.

fixes #1664