feat(evm): save a syscall for most precompile calls
Previously, we'd query the remaining gas for every precompile call. Now,
we only query the remaining gas when actually making a real call (i.e.,
only in the "native call" precompile.

Unfortunately, we make a syscall to query gas because we don't yet have
a way to directly access this information from wasm. Ideally we'd implement
https://github.com/filecoin-project/FIPs/discussions/845 and avoid this
issue entirely, but that's a much larger project.

This should save ~14k gas per pre-compile call, which isn't nothing.